requirements.txt
matplotlib==3.10.1
mysql_connector==0.3.1
pandas==2.2.3
seaborn==0.13.2
setuptools==65.5.0
setuptools==68.2.0
SQLAlchemy==2.0.37
streamlit==1.44.1
